Stop Bot Incorrectly Accusing People of Spamming.
The bot often incorrectly accuses people of spamming. There has never been an instance of an actual spammer spamming mass messages to my recorlection, so I have doubled the amount of messages and interval it takes before the warning and ban kicks in. I have also changed the warning message to be more appropriate and kinder.